+++ src/DependencyInjection/AuditTrailExtension.php
@@ -4,13 +4,16 @@ declare(strict_types=1);
 
 namespace Rcsofttech\AuditTrailBundle\DependencyInjection;
 
+use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractCrudController;
 use LogicException;
 use Override;
 use Rcsofttech\AuditTrailBundle\Contract\AuditTransportInterface;
+use Rcsofttech\AuditTrailBundle\Controller\Admin\AuditLogCrudController;
 use Rcsofttech\AuditTrailBundle\DataCollector\AuditDataCollector;
 use Rcsofttech\AuditTrailBundle\DataCollector\TraceableAuditCollector;
 use Rcsofttech\AuditTrailBundle\Factory\AuditLogMessageFactory;
 use Rcsofttech\AuditTrailBundle\MessageHandler\PersistAuditLogHandler;
+use Rcsofttech\AuditTrailBundle\Serializer\AuditLogMessageSerializer;
 use Rcsofttech\AuditTrailBundle\Transport\AsyncDatabaseAuditTransport;
 use Rcsofttech\AuditTrailBundle\Transport\ChainAuditTransport;
 use Rcsofttech\AuditTrailBundle\Transport\DoctrineAuditTransport;
@@ -32,10 +35,6 @@ use function count;
 
 final class AuditTrailExtension extends Extension implements PrependExtensionInterface
 {
-    private const string EASYADMIN_CRUD_CONTROLLER = 'Rcsofttech\\AuditTrailBundle\\Controller\\Admin\\AuditLogCrudController';
-
-    private const string EASYADMIN_BASE_CRUD_CONTROLLER = 'EasyCorp\\Bundle\\EasyAdminBundle\\Controller\\AbstractCrudController';
-
     #[Override]
     public function load(array $configs, ContainerBuilder $container): void
     {
@@ -100,12 +99,16 @@ final class AuditTrailExtension extends Extension implements PrependExtensionInt
         $container->setParameter('audit_trail.integrity.algorithm', $config['integrity']['algorithm']);
         new YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'))->load('services.yaml');
 
+        if (interface_exists(MessageBusInterface::class)) {
+            $this->registerMessengerSupport($container);
+        }
+
         $this->configureTransports($config, $container);
 
         /** @var array<string, mixed> $bundles */
         $bundles = $container->hasParameter('kernel.bundles') ? $container->getParameter('kernel.bundles') : [];
-        if (isset($bundles['EasyAdminBundle']) && class_exists(self::EASYADMIN_BASE_CRUD_CONTROLLER)) {
-            $container->register(self::EASYADMIN_CRUD_CONTROLLER, self::EASYADMIN_CRUD_CONTROLLER)
+        if (isset($bundles['EasyAdminBundle']) && class_exists(AbstractCrudController::class)) {
+            $container->register(AuditLogCrudController::class, AuditLogCrudController::class)
                 ->setAutowired(true)
                 ->setAutoconfigured(true)
                 ->setArgument('$adminPermission', '%audit_trail.admin_permission%')
@@ -189,8 +192,6 @@ final class AuditTrailExtension extends Extension implements PrependExtensionInt
             throw new LogicException('To use async database transport, you must install the symfony/messenger package.');
         }
 
-        $this->registerMessageFactory($container);
-
         $handlerId = 'rcsofttech_audit_trail.handler.persist_audit_log';
         $container->register($handlerId, PersistAuditLogHandler::class)
             ->setAutowired(true)
@@ -211,6 +212,18 @@ final class AuditTrailExtension extends Extension implements PrependExtensionInt
         }
 
         $container->register(AuditLogMessageFactory::class, AuditLogMessageFactory::class)
+            ->setAutowired(true);
+    }
+
+    private function registerMessengerSupport(ContainerBuilder $container): void
+    {
+        $this->registerMessageFactory($container);
+
+        if ($container->has(AuditLogMessageSerializer::class)) {
+            return;
+        }
+
+        $container->register(AuditLogMessageSerializer::class, AuditLogMessageSerializer::class)
             ->setAutowired(true);
     }
 
@@ -242,8 +255,6 @@ final class AuditTrailExtension extends Extension implements PrependExtensionInt
         if (!interface_exists(MessageBusInterface::class)) {
             throw new LogicException('To use the Queue transport, you must install the symfony/messenger package.');
         }
-
-        $this->registerMessageFactory($container);
 
         $id = 'rcsofttech_audit_trail.transport.queue';
         $definition = $container->register($id, QueueAuditTransport::class)